requirements.txt
fastapi~=0.103
uvicorn[standard]
python-dotenv~=1.0.0